Create digital watermark
Due to the simplicity of picture tampering, ensuring the authenticity of digital photographs has grown to be a crucial concern in the age of digital transformation. Researchers have been actively tackling this problem for the past few decades, attempting to create efficient image watermarking methods suited to varied applications. The challenge of developing a watermarking system that balances security and resilience still exists. This article summarizes common watermarking frameworks and identifies the important criteria that need be met before developing new techniques for various purposes. It also examines contemporary developments in digital picture watermarking with the purpose of identifying novel approaches and their accompanying drawbacks. Additionally, the report examines conventional attacks and offers prospective lines of inquiry for further study.This research used a novel blind A digital picture watermarking technique that functions well with both color and grayscale photos is introduced. The method uses the Discrete Cosine Transform (DCT) as a first step before embedding the watermark. Before performing the DCT, the host image is partitioned into 8x8 nonoverlapping blocks, and the watermark bit is integrated by adjusting the difference between the DCT coefficients of neighboring blocks. The Arnold transform and chaotic encryption are coupled to increase security, resulting in a two-layer security strategy for the watermark. The findings from simulations show that the suggested method demonstrates resilience to various image processing operations such as compression, sharpening, cropping, and median filtering. The paper studies and evaluates three alternative iterations of this algorithm. In a comparison with current methods, the suggested scheme performs better than average in terms of imperceptibility, security, and robustness. Given these benefits, the suggested technique shows promise for use in fields like telemedicine and e-healthcare, providing a reliable solution for hiding electronic health records within medical photographs.